doubly linked list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -1,12 +1,12 @@
 #include "lists.h"
 /**
  * print_dlistint - prints all the elements of a list_t list
- * @h: head of the list
+ * @h: pointer to head of the list
  * Return: number of nodes printed
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	unsigned int counter;
+	size_t counter;
 
 	if (h == NULL)
 		return (0);
